LPS-101720 Fix typo 'contants' to 'constants' in package path
