+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-checker-email-otp-web/src/main/java/com/liferay/multi/factor/authentication/checker/email/otp/web/internal/portlet/action/MFAVerifyMVCRenderCommand.java
@@ -15,8 +15,8 @@
 package com.liferay.multi.factor.authentication.checker.email.otp.web.internal.portlet.action;
 
 import com.liferay.multi.factor.authentication.checker.email.otp.web.internal.checker.EmailOTPMFAChecker;
-import com.liferay.multi.factor.authentication.checker.email.otp.web.internal.contants.MFAPortletKeys;
-import com.liferay.multi.factor.authentication.checker.email.otp.web.internal.contants.MFAPortletURLFactory;
+import com.liferay.multi.factor.authentication.checker.email.otp.web.internal.constants.MFAPortletKeys;
+import com.liferay.multi.factor.authentication.checker.email.otp.web.internal.constants.MFAPortletURLFactory;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
